move _getSnapshots
